Fix Scala compatibility
